+++ packages/headless/src/controllers/insight/quickview/headless-insight-quickview.ts
@@ -1,31 +1,28 @@
 import {InsightEngine} from '../../../app/insight-engine/insight-engine';
 import {logDocumentQuickview} from '../../../features/result-preview/result-preview-analytics-actions';
 import {
-  QuickviewCoreProps,
-  QuickviewCoreOptions,
-  QuickviewCore,
-  QuickviewCoreState,
-  buildQuickviewCore,
+  QuickviewProps,
+  QuickviewOptions,
+  Quickview,
+  QuickviewState,
+  buildCoreQuickview,
 } from '../../core/quickview/headless-core-quickview';
 
-export interface InsightQuickviewProps extends QuickviewCoreProps {}
-export interface InsightQuickviewOptions extends QuickviewCoreOptions {}
-export interface InsightQuickview extends QuickviewCore {}
-export interface InsightQuickviewState extends QuickviewCoreState {}
+export type {QuickviewOptions, QuickviewState, QuickviewProps, Quickview};
 
 /**
- * Creates an `InsightQuickview` controller instance.
- * @param engine - The headless engine.
- * @param props - The configurable `InsightQuickview` properties.
- * @returns An `InsightQuickview` controller instance.
+ * Creates an insight `Quickview` controller instance.
+ * @param engine - The insight engine.
+ * @param props - The configurable `Quickview` properties.
+ * @returns A `Quickview` controller instance.
  */
-export function buildInsightQuickview(
+export function buildQuickview(
   engine: InsightEngine,
-  props: InsightQuickviewProps
-): InsightQuickview {
+  props: QuickviewProps
+): Quickview {
   const fetchResultContentCallback = () => {
     engine.dispatch(logDocumentQuickview(props.options.result));
   };
 
-  return buildQuickviewCore(engine, props, fetchResultContentCallback);
+  return buildCoreQuickview(engine, props, fetchResultContentCallback);
 }